✨feat: 新增校验函数，获取数据类型函数,整理文档
