--- docs/utils-docs/RelativeOrAbsolute/index.md
@@ -1,9 +1,0 @@
----
-title: relative-absolute
-group:
-  title: 数据转换
----
-
-# 相对路径 转 绝对路径
-
-<code src="./RelativeToAbsolute.tsx">相对路径 转 绝对路径</code>
