--- docs/utils-docs/RgbaOrHex/HexToRgba.tsx
@@ -1,5 +1,0 @@
-import RgbaHex from './RgbaHex';
-
-const HexToRgba: React.FC = () => <RgbaHex initial={'#66ccff'} inputType={'hex'} />;
-
-export default HexToRgba;
