--- docs/utils-docs/RgbaOrHex/RgbaToHex.tsx
@@ -1,5 +1,0 @@
-import RgbaOrHex from './RgbaHex';
-
-const RgbaToHex: React.FC = () => <RgbaOrHex initial={'rgba(102,204,255,1)'} inputType={'rgba'} />;
-
-export default RgbaToHex;
