--- docs/utils-docs/RgbaOrHex/index.md
@@ -1,13 +1,0 @@
----
-title: rgba-hex
-group:
-  title: 数据转换
----
-
-# rgba 转 16 进制
-
-<code src="./RgbaToHex.tsx">rgba 转 16 进制</code>
-
-# 16 进制 转 rgba
-
-<code src="./HexToRgba.tsx">16 进制 转 rgba</code>
