+++ docs/utils-docs/crypto-aes/index.md
@@ -1,0 +1,11 @@
+---
+title: CryptoAES
+nav:
+  title: 无状态函数
+group:
+  title: 数据加密
+---
+
+# CryptoAES
+
+<code src="./CryptoAES.tsx">AES</code>
