+++ docs/utils-docs/crypto-base64/index.md
@@ -1,0 +1,11 @@
+---
+title: CryptoBase64
+nav:
+  title: 无状态函数
+group:
+  title: 数据加密
+---
+
+# CryptoBase64
+
+<code src="./CryptoBase64.tsx">Base64</code>
