--- docs/utils-docs/crypto/index.md
@@ -1,15 +1,0 @@
----
-title: crypto
-nav:
-  title: 无状态函数
-group:
-  title: 数据转换
----
-
-# Base64
-
-<code src="./Base64.tsx">Base64</code>
-
-# AES
-
-<code src="./AES.tsx">AES</code>
