+++ docs/utils-docs/get-absolute-path/index.md
@@ -1,0 +1,9 @@
+---
+title: getAbsolutePath
+group:
+  title: get函数合集
+---
+
+# 获取绝对路径
+
+<code src="./getAbsolutePath.tsx">获取绝对路径</code>
