--- docs/utils-docs/htmlOrString/index.md
@@ -1,13 +1,0 @@
----
-title: html-string
-group:
-  title: 数据转换
----
-
-# 实体字符转字符串
-
-<code src="./HtmlToString.tsx">实体字符转字符串</code>
-
-# 字符串转实体字符
-
-<code src="./StringToHtml.tsx">字符串转实体字符</code>
