--- docs/utils-docs/styledToString/index.md
@@ -1,9 +1,0 @@
----
-title: styled-string
-group:
-  title: 数据转换
----
-
-# 将styled对象转为css-in-js可以运行的字符串
-
-<code src="./StyledToString.tsx">styled 转成 string</code>
