+++ docs/utils-docs/transform-3-html-to-string/index.md
@@ -1,0 +1,10 @@
+---
+title: htmlToString
+order: 3
+group:
+  title: 数据转换
+---
+
+# 实体字符转字符串
+
+<code src="./HtmlToString.tsx">实体字符转字符串</code>
