+++ docs/utils-docs/transform-4-string-to-html/index.md
@@ -1,0 +1,10 @@
+---
+title: stringToHtml
+order: 3
+group:
+  title: 数据转换
+---
+
+# 字符串转实体字符
+
+<code src="./StringToHtml.tsx">字符串转实体字符</code>
