+++ docs/utils-docs/transform-5-rgba-to-hex/index.md
@@ -1,0 +1,10 @@
+---
+title: rgbaToHex
+order: 5
+group:
+  title: 数据转换
+---
+
+# rgba 转 16 进制
+
+<code src="./RgbaToHex.tsx">rgba 转 16 进制</code>
