+++ docs/utils-docs/transform-6-hex-to-rgba/index.md
@@ -1,0 +1,10 @@
+---
+title: hexToRgba
+order: 6
+group:
+  title: 数据转换
+---
+
+# 16 进制 转 rgba
+
+<code src="./HexToRgba.tsx">16 进制 转 rgba</code>
