+++ docs/utils-docs/transform-7-styled-to-string/index.md
@@ -1,0 +1,10 @@
+---
+title: styled-string
+order: 7
+group:
+  title: 数据转换
+---
+
+# 将styled对象转为css-in-js可以运行的字符串
+
+<code src="./StyledToString.tsx">styled 转成 string</code>
