+++ package.json
@@ -1,7 +1,7 @@
 {
   "name": "@dxsixpc/utils",
-  "version": "0.2.2",
-  "description": "工具库",
+  "version": "1.0.0-alpha.1",
+  "description": "工具函数库",
   "homepage": "https://dxsixpc.github.io/utils",
   "repository": "https://github.com/dxsixpc/utils",
   "license": "MIT",
@@ -58,12 +58,12 @@
     "@dxsixpc/configs": "^1.0.3",
     "@emotion/react": "^11.10.6",
     "@emotion/styled": "^11.10.6",
-    "@types/jest": "^29.5.0",
+    "@types/jest": "^29.5.1",
     "@types/lodash": "^4.14.194",
     "@types/node": "^18.15.11",
-    "@types/react": "^18.0.35",
+    "@types/react": "^18.0.37",
     "@types/react-dom": "^18.0.11",
-    "antd": "^5.4.2",
+    "antd": "^5.4.3",
     "babel-jest": "^29.5.0",
     "clipboard": "^2.0.11",
     "commitizen": "^4.0.3",
@@ -87,10 +87,6 @@
     "ts-node": "^10.9.1",
     "typescript": "^5.0.4",
     "webpack": "^5.79.0"
-  },
-  "peerDependencies": {
-    "react": ">=16.9.0",
-    "react-dom": ">=16.9.0"
   },
   "publishConfig": {
     "access": "public"
