+++ src/helpers/request/__tests__/index.test.ts
@@ -1,13 +1,15 @@
 import request from '../request';
 
-test('request:success', async () => {
-  await request.get('https://www.baidu.com/').then((res) => {
-    expect(res.status).toBe(200);
+describe('请求接口', () => {
+  test('成功', async () => {
+    await request.get('https://www.baidu.com/').then((res) => {
+      expect(res.status).toBe(200);
+    });
   });
-});
 
-test('request:error', async () => {
-  await request.get('https://miao.baidu.com/abdr').catch((err) => {
-    expect(err.status).toBe(400);
+  test('失败', async () => {
+    await request.get('https://miao.baidu.com/abdr').catch((err) => {
+      expect(err.status).toBe(400);
+    });
   });
 });
