+++ src/utils/getAbsolutePath/__tests__/index.test.ts
@@ -1,0 +1,8 @@
+/**
+ * @jest-environment jsdom
+ */
+import relativeToAbsolute from '../getAbsolutePath';
+
+test('相对路径转绝对路径', () => {
+  expect(relativeToAbsolute('/utils/logo.png')).toBe('http://localhost/utils/logo.png');
+});
