+++ src/utils/getAbsolutePath/index.ts
@@ -1,0 +1,1 @@
+export { default as getAbsolutePath } from './getAbsolutePath';
