+++ src/utils/getType/index.ts
@@ -1,0 +1,2 @@
+export { default as getType } from './getType';
+export type { DataType } from './getType';
