+++ src/utils/index.ts
@@ -1,7 +1,9 @@
 // utils文件夹里，是一些简单的无状态的静态工具函数；
 export * from './crypto';
 export * from './dataOrString';
+export * from './getAbsolutePath';
+export * from './getType';
 export * from './htmlOrString';
-export * from './relativeOrAbsolute';
 export * from './rgbaOrHex';
 export * from './styledToString';
+export * from './validate';
