--- src/utils/relativeOrAbsolute/index.ts
@@ -1,1 +1,0 @@
-export { default as relativeToAbsolute } from './relativeToAbsolute';
