+++ src/utils/validate/index.ts
@@ -1,0 +1,6 @@
+export { default as validate } from './validate';
+export type { ValidateType } from './validate';
+export { default as validateEmail } from './validateEmail';
+export { default as validateIDCard } from './validateIDCard';
+export { default as validatePhone } from './validatePhone';
+export { default as validateUrl } from './validateUrl';
