+++ src/utils/validate/validateDate.ts
@@ -1,0 +1,8 @@
+/**
+ * @name 验证日期
+ */
+const validateDate = (data: string) => {
+  return !Number.isNaN(new Date(data).valueOf());
+};
+
+export default validateDate;
