+++ src/utils/validate/validateIDCard.ts
@@ -1,0 +1,10 @@
+/**
+ * @name 验证身份证
+ */
+const validateIDCard = (data: string) => {
+  return /^[1-9]\d{7}((0\d)|(1[0-2]))(([0|1|2]\d)|3[0-1])\d{3}$|^[1-9]\d{5}[1-9]\d{3}((0\d)|(1[0-2]))(([0|1|2]\d)|3[0-1])\d{3}([0-9]|X)$/.test(
+    data,
+  );
+};
+
+export default validateIDCard;
