+++ src/utils/validate/validateUrl.ts
@@ -1,0 +1,8 @@
+/**
+ * @name 验证Url链接
+ */
+const validateUrl = (data: string) => {
+  return /^(http|https|ftp|file):\/\/([\w.]+\/?)\S*/.test(data);
+};
+
+export default validateUrl;
